Split de  command into several files
